Build/source updates
* Updates to docstrings and error messages
* Add DIY directories to CFLAGS and LDFLAGS
* Improvements to gdb for reactiviting the scheduler on 'c'
* Add more commands for using sessions in gdb
* Add some paranoid, compiler-calming NULL pointer checks
* Have configure do a better job of identifying when to use lib64
* Add KNO_CHECK_XTYPE_OUTLEN config
* Simplify .gitmodules
* Use macro to speed up ascii u8_base_char
* Add more U8_INLINE_IO declarations
* Added precommit hook
* Move extmods to contribs, add hunspell
* Add smarter knogdb which can just take a PID
* Add more gdb custom commands, especially for sessions
* Added initial importdata contrib
* Add DIY build flags

Module updates

* Add fileinfo stringfmts
* Got defstruct to work, using it with fifo to test
* Added stubs for workpools and knodb queries

KNOC changes

* Fix initsession config get handling
* Have debugging switch to new errors, and add more default session commands
* Add initial help/apropos module
* Rely on default args to session-config
* Add kno/help module to appmods for KNOC
* Improve value summaries for the help interface
* Get session=none config to work and added LINKSESSION config
* Make the value history more liberal
* Make ,= assignment give more information for choice assignments

Storage changes

* Add optional tableops to props/metadata dbctl operations
* Add appendix handling and more indexloc specification options
* Provide distinguishing id for upstart pool on conflicts
* Warn when an adjunct can't be opened
* Handle empty db arguments for prefetch primitives
* Change hashtable_keyvals to be able to support different commit flows
* Clean up some symbols and add more structure/detail to dbctl stats operation
* Fix code for updating pool loads
* Add dbctl tests for unsupported operations
* Don't bother calling commit functions when there's nothing to do
* Fix argument handling for index/save!
* Fixes to kindex bucket functions
* Add knodb stats command
* Fix pool display to use pool-label
* Handle swapout with ambiguous pool args
* Switch back to having default open-index/pool look up existing indexes
* Fix bug in shared pool/index open, pass shared flags alongside register #f

* KNODB updates

* Better backup handling, reorganize source
* Add name/name.pool handling to db configfns
* Improvements to db packing commands
* Simplify and refactor knodb/config
* Add better pool index specification, fix some knodb/readonly issues
* Handle branching of flexindexes
* Add reserved props to flexindex partitions to help even out allocation
* Have 'required option for backup functions to not allow simple deletion
* Add knodb/alldbs
* Add quiet flag for logctl and use it in knodb command
* Improve db config conflict resolution
* Rename splitindex to copykeys
* Many fixes/changes to flexpack and related operations/modules
* Improve db configs to handle different suffixes, and have knodb/same-path?
* Add metadata poolctl handler for flexpools
* Add aggregate-index? case to get-modified (and callers)
* Add hashindex/slotcounts
* Add flexindex/ref to support unique named flexindexes
* Export dbref resolver (knodb/config-dbref) from knodb/config
* Use flexindex/ref to resolve flexindexes
* Flex fixes, including thread locking bug in extending flexindexes
* Have stderr output of flexpack subprocs be temporary
* Add more smarts to automatic repacking
* Improve DBConflict log diagnostics

Indexing updates

* Add fuzzing (knodb/fuzz) implementation
* Improvements to flexindexes: better reports, jobids, allow identification of multiple fronts
* Have flexindex report total run time, provide more informative jobids
* Fix some issues in index merge/save
* Add change load information to db commit messages
* Have branch/commit able to use index_save as well as index_merge
* Improvements to knodb/index functions, including fuzzing
* Add flags/opts for controlling rollback execution, don't use open un-registered indexes as registered
* Add stdpunct primitive with tests
* Add stopwords/stopchecks to text indexing
* Add indexing metadata to pools
* Add index-keyslots primitive to get or test index keyslots
* Add better index stats, provide option to avoid rollbacks
* Add options for avoiding rollbacks
* Add readonly indexctl to aggregates
* Copy base metadata when packing indexes
* Sort flexindex partitions by size (large to small) when merging
* Added knodb/fuzz/graph
* Do direct branch merges when repacking hashindexes
* Add %wrapX forms to support debugging
* Clean up any leftover values in sloppy use of locals
* Add %weval to %wrap to optimize, renamed some leftover commentedion %watch statements
* Fix leak in do_difference, leak/duplicate call in d1_call invocation
* Added typeinfo aliases
* Fix fatal recursion in dbct/type dispatch interaction, add pool/index aliases for kno_pool_type and kno_index_type and use them for dumping pools as xtypes
* Don't use readonly index specs for targets
* Have flexindexes only include numbered partitions by default

Engine updates

* Improvements to engines: stopval fixes, fifo updates, debug levels, commit load check functions,
* Tweak engine debug levels
* Work towards checkpoint functions based on change load, add asynchronous direct-to-disk branch commits
* Don't store threads on loop-data to avoid leak, provide threadhook option to be called on threads externally
* Have engine/delta test return an explanatory result
* Have engine report interval tests symbolically
* Update fifo fill logging and improve bugjar logging
* Add branchmerge engine option to allow direct index/save of branches, add checksync engine option to pause when checkpoints take too long or lock during checkpoints
* Handle #f arguments in fifo making
* Make  bytes/sec be smarter for large rates
* Add optional posfn to engine/readfile enabling results to be of the form (pos . data)
* Remove old bulk commit model from engine, now handled by knodb
* Don't assume a savefreq if there are maxchanges specified

Runtime changes

* Pick up u8run environment variables
* Fixed bug with output of negative separated fixnums
* Have testopt be an ndcall handling ambiguous keys
* Use contextual detail args in listing embedded tables
* Fix loglevel configuration for subprocs
* Use main.setup for the setup method
* Added config probes
* Add app setup to knox
* Have exit actually call exit(), have proc/open files opened for append
* make config parsing more robust
* Add options arg to mt/call
* Add shorter aliases for ctime/mtime/atime functions
* Change histref syntax to use embedded # or , separators, rather than ., fixed some code style issues
* Update config system to use tables rather than linear lookup on a linked list, closes #163
* Added global subprocs list, closes #136
* Output sourcebase in log messages when set
* Add KNO_SUPPRESS_MESSAGES environment variable to init kno_be_vewy_quiet
* Started adding extended KNO_PROMISE functionality

Optimizer updates

* Added env-module to get the top level module for an environment
* Fix bug in tryif optimizer
* Use env-module for recording warnings
* logmsg updates

+++ src/prims/procprims.c
@@ -11,6 +11,7 @@
 #include "kno/knosource.h"
 #include "kno/lisp.h"
 #include "kno/numbers.h"
+#include "kno/tables.h"
 #include "kno/apply.h"
 #include "kno/storage.h"
 #include "kno/pools.h"
@@ -119,7 +120,7 @@ static int stringvec_len(char **vec)
   return scan-vec;
 }
 
-static int default_subproc_loglevel = LOG_WARN;
+static int default_subproc_loglevel = LOG_INFO;
 
 #define ARG_ESCAPE_SCHEME 1
 #define ARG_ESCAPE_CONFIGS 2
@@ -204,7 +205,9 @@ static void finish_subproc(struct KNO_SUBPROC *p)
 	else errfilename = KNO_CSTRING(err);}
       else if (KNO_TYPEP(err,kno_ioport_type)) {}
       else NO_ELSE;
-      int loglevel  = (normal_exit) ? (p->proc_loglevel) : (LOGERR) ;
+      int loglevel  = (normal_exit) ? 
+	( (p->proc_loglevel>=0) ? (p->proc_loglevel) : (default_subproc_loglevel) ) :
+	(LOGERR);
       if (WIFEXITED(status))
 	u8_log(loglevel,"SubprocFinished","PID=%lld %s status=%d %s%s%s%s",
 	       (long long)(p->proc_pid),p->proc_id,WEXITSTATUS(status),
@@ -686,7 +689,8 @@ static lispval proc_run_prim(int n,kno_argvec args)
   else if (KNO_STRINGP(outfile)) {
     u8_string use_file = u8_abspath(KNO_CSTRING(outfile),out_dir);
     kno_decref(outfile); outfile = kno_init_string(NULL,-1,use_file);
-    int new_stdout = u8_open_fd(use_file,O_WRONLY|O_CREAT,STDOUT_FILE_MODE);
+    int new_stdout = u8_open_fd
+      (use_file,O_WRONLY|O_APPEND|O_CREAT,STDOUT_FILE_MODE);
     if (new_stdout<0) {
       u8_exception ex = u8_current_exception;
       u8_log(LOGCRIT,"Couldn't open %s as stdout (%s:%s)",
@@ -709,7 +713,8 @@ static lispval proc_run_prim(int n,kno_argvec args)
   else if (KNO_STRINGP(errfile)) {
     u8_string use_file = u8_abspath(KNO_CSTRING(errfile),out_dir);
     kno_decref(errfile); errfile = kno_init_string(NULL,-1,use_file);
-    int new_stderr = u8_open_fd(use_file,O_WRONLY|O_CREAT,STDERR_FILE_MODE);
+    int new_stderr =
+      u8_open_fd(use_file,O_WRONLY|O_APPEND|O_CREAT,STDERR_FILE_MODE);
     if (new_stderr<0) {
       u8_exception ex = u8_current_exception;
       u8_log(LOGCRIT,"Couldn't open %s as stderr (%s:%s)",
@@ -731,7 +736,7 @@ static lispval proc_run_prim(int n,kno_argvec args)
     else err = errpipe;}
 
   lispval wait_opt = kno_getopt(opts,KNOSYM(wait),KNO_FALSE);
-  int nofork = flags&(KNO_NO_FORK);
+   int nofork = flags&(KNO_NO_FORK);
   pid_t pid = (nofork) ? (0) : (fork());
 
   /* Once we have a PID, we can generate stdout/stderr files with the
@@ -808,7 +813,7 @@ static lispval proc_run_prim(int n,kno_argvec args)
       int rv = waitpid(pid,&status,0);
       if (rv>=0) subproc->proc_status = status;
       if (rv<0) {
-	lispval err = kno_err("WaitFailed","proc_run",progname,result);
+	lispval err = kno_err("WaitFailed","proc_run",subproc->proc_id,result);
 	kno_decref(result);
 	result=err;}
       else if ( (WIFEXITED(status)) || (WIFSIGNALED(status)) )
@@ -817,7 +822,8 @@ static lispval proc_run_prim(int n,kno_argvec args)
     errno = 0;}
   else {
     int err = errno; errno=0;
-    result = kno_err(u8_strerror(err),"proc_run/wait",progname,KNO_VOID);}
+    result = kno_err(u8_strerror(err),"proc_run/wait",idstring,KNO_VOID);
+    u8_free(idstring); idstring=NULL;}
 
   /* Clean up any sockets we created */
   /* pair[0] is the read end, pair[1] is the write end */
@@ -843,7 +849,8 @@ static lispval proc_run_prim(int n,kno_argvec args)
   if (KNO_ABORTED(result)) {
     if (cprogname) u8_free(cprogname);}
   if (out_dir) u8_free(out_dir);
-  if (idbase) u8_free(idbase); idbase=NULL;
+  if (idbase) u8_free(idbase);
+  idbase=NULL;
   /* free_stringvec(argv); free_stringvec(envp); */
   kno_decref(infile);
   kno_decref(outfile);
@@ -1201,6 +1208,9 @@ DEFC_PRIM("exit",exit_prim,
 static lispval exit_prim(lispval arg)
 {
   pid_t main_thread = getpid();
+  if (KNO_FIXNUMP(arg)) {
+    int val = KNO_FIX2INT(arg);
+    exit(val);}
   int rv = kill(main_thread,SIGTERM);
   if (rv<0)
     return kno_err("ExitFailed","exit_prim",NULL,VOID);
@@ -1505,6 +1515,14 @@ static lispval proc_stoppedp(lispval arg)
   else return kno_err("NotAProc","proc/running?",NULL,arg);
 }
 
+DEFC_PRIM("all-subprocs",all_subprocs,
+	  KNO_MAX_ARGS(0)|KNO_MIN_ARGS(0),
+	  "Returns a choice of all recorded subpcos")
+static lispval all_subprocs()
+{
+  return kno_hashtable_values(proctable);
+}
+
 /* SUBPROC object */
 
 static lispval cons_subproc(pid_t pid,u8_string id,
@@ -1522,7 +1540,7 @@ static lispval cons_subproc(pid_t pid,u8_string id,
   proc->proc_id = id;
   proc->proc_pid = pid;
   proc->proc_status = 0;
-  proc->proc_loglevel = -1;
+  proc->proc_loglevel = loglevel;
   proc->proc_progname = cprogname;
   proc->proc_argv = argv;
   proc->proc_envp = envp;
@@ -1645,7 +1663,7 @@ KNO_EXPORT void kno_init_procprims_c()
 
 static void link_local_cprims()
 {
-  KNO_LINK_CPRIMN("proc/run",proc_run_prim,procprims_module);
+  KNO_LINK_CPRIMN("%proc/run",proc_run_prim,procprims_module);
   KNO_LINK_CPRIM("proc-pid",proc_pid,1,procprims_module);
   KNO_LINK_CPRIM("proc-id",proc_idstring,1,procprims_module);
   KNO_LINK_CPRIM("proc-progname",proc_progname,1,procprims_module);
@@ -1676,4 +1694,6 @@ static void link_local_cprims()
   KNO_LINK_CPRIM("proc/status",proc_status,1,procprims_module);
   KNO_LINK_CPRIM("proc/info",proc_info,1,procprims_module);
   KNO_LINK_CPRIM("proc/wait",proc_wait,1,procprims_module);
+
+  KNO_LINK_CPRIM("all-subprocs",all_subprocs,0,procprims_module);
 }
